Upgrade terraform provider alicloud to v1.244.0 (#946)
This PR was generated via $ upgrade-provider pulumi/pulumi-alicloud
--kind=provider.

Upgrading terraform-provider-alicloud
Fixes https://github.com/pulumi/pulumi-alicloud/issues/937

+++ sdk/dotnet/SimpleApplicationServer/Outputs/GetServerPlansPlanResult.cs
@@ -36,7 +36,7 @@ namespace Pulumi.AliCloud.SimpleApplicationServer.Outputs
         /// <summary>
         /// The memory size. Unit: GB.
         /// </summary>
-        public readonly int Memory;
+        public readonly double Memory;
         /// <summary>
         /// The ID of the Instance Plan.
         /// </summary>
@@ -58,7 +58,7 @@ namespace Pulumi.AliCloud.SimpleApplicationServer.Outputs
 
             string id,
 
-            int memory,
+            double memory,
 
             string planId,
